allow real js files for stubs
