+++ packages/renderer/element-js/index.js
@@ -1,21 +1,12 @@
 import ElementFactory from './element-factory';
 import TemplateRenderer from "./template-renderer";
 
-const stubs = [{
+const stubs = async () => [{
     sources: ['@webtides/element-js'],
-    stub: `
-        import { html } from '@popeindustries/lit-html-server';
-        class TemplateElement {}
-        class StyledElement {}
-        class BaseElement {}
-        export { TemplateElement, StyledElement, BaseElement, html }
-    `
+    stub: require.resolve('./stubs/index.js'),
 }, {
     sources: ['lit-html/directives/unsafe-html'],
-    stub: `
-        import { unsafeHTML } from '@popeindustries/lit-html-server/directives/unsafe-html';
-        export { unsafeHTML }
-    `
+    stub: require.resolve('./stubs/directives/unsafe-html.js'),
 }];
 
 export {ElementFactory, TemplateRenderer, stubs};
