+++ packages/renderer/element-js/stubs/directives/unsafe-html.js
@@ -1,0 +1,2 @@
+import { unsafeHTML } from '@popeindustries/lit-html-server/directives/unsafe-html';
+export { unsafeHTML }
