+++ packages/renderer/element-js/stubs/index.js
@@ -1,0 +1,7 @@
+import { html } from '@popeindustries/lit-html-server';
+
+class TemplateElement {}
+class StyledElement {}
+class BaseElement {}
+
+export { TemplateElement, StyledElement, BaseElement, html }
